# mvp/warehouse-service/swagger.yaml
openapi: 3.0.0
info:
  title: Warehouse Service API
  description: API для сервиса управления складом и учета оборудования
  version: 1.0.0
servers:
  - url: http://localhost:8001
    description: Локальный сервер разработки
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateItemRequest:
      type: object
      required:
        - name
        - serial_number
        - category
        - quantity
        - location
      properties:
        name:
          type: string
          description: Название оборудования
        serial_number:
          type: string
          description: Серийный номер оборудования
        category:
          type: string
          description: Категория оборудования
        description:
          type: string
          description: Описание оборудования
        manufacturer:
          type: string
          description: Производитель оборудования
        price:
          type: number
          format: float
          description: Стоимость единицы оборудования
        quantity:
          type: integer
          description: Количество единиц оборудования
        min_quantity:
          type: integer
          description: Минимальное количество единиц оборудования на складе
        location:
          type: string
          description: Местоположение оборудования на складе
        purchase_date:
          type: string
          format: date-time
          description: Дата покупки оборудования
        warranty_expiry:
          type: string
          format: date-time
          description: Дата окончания гарантии
        status:
          type: string
          enum: [available, reserved, unavailable]
          default: available
          description: Статус оборудования на складе
    UpdateItemRequest:
      type: object
      properties:
        name:
          type: string
          description: Название оборудования
        category:
          type: string
          description: Категория оборудования
        description:
          type: string
          description: Описание оборудования
        manufacturer:
          type: string
          description: Производитель оборудования
        quantity:
          type: integer
          description: Количество единиц оборудования
        min_quantity:
          type: integer
          description: Минимальное количество единиц оборудования на складе
        location:
          type: string
          description: Местоположение оборудования на складе
        purchase_date:
          type: string
          format: date-time
          description: Дата покупки оборудования
        warranty_expiry:
          type: string
          format: date-time
          description: Дата окончания гарантии
        status:
          type: string
          enum: [available, reserved, unavailable]
          description: Статус оборудования на складе
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Текст ошибки
        status:
          type: integer
          description: HTTP статус код
...
paths:
...
  /items:
    post:
      summary: Создание нового оборудования на складе
      description: Добавляет новое оборудование в базу данных склада и создает начальную транзакцию прихода
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemRequest"
      responses:
        "201":
          description: Оборудование успешно создано
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  item:
                    $ref: "#/components/schemas/WarehouseItem"
        "400":
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Оборудование с таким серийным номером уже существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      summary: Получение списка оборудования
      description: Возвращает список оборудования на складе с возможностью фильтрации
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Ограничение количества результатов
          schema:
            type: integer
            default: 50
        - name: skip
          in: query
          description: Количество пропускаемых результатов (для пагинации)
          schema:
            type: integer
            default: 0
        - name: category
          in: query
          description: Фильтр по категории оборудования
          schema:
            type: string
        - name: location
          in: query
          description: Фильтр по местоположению
          schema:
            type: string
        - name: status
          in: query
          description: Фильтр по статусу
          schema:
            type: string
            enum: [available, reserved, unavailable]
        - name: sort
          in: query
          description: Поле для сортировки
          schema:
            type: string
            default: created_at
        - name: search
          in: query
          description: Поисковый запрос (по названию или серийному номеру)
          schema:
            type: string
      responses:
        "200":
          description: Список оборудования
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    description: Общее количество оборудования
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/WarehouseItem"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /items/{id}:
    get:
      summary: Получение информации об оборудовании
      description: Возвращает детальную информацию о конкретном оборудовании
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор оборудования
          schema:
            type: string
      responses:
        "200":
          description: Информация об оборудовании
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseItem"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Оборудование не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Обновление информации об оборудовании
      description: Обновляет данные о конкретном оборудовании
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор оборудования
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemRequest"
      responses:
        "200":
          description: Оборудование успешно обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseItem"
        "400":
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Оборудование не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Удаление оборудования со склада
      description: Удаляет оборудование из базы данных склада
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор оборудования
          schema:
            type: string
      responses:
        "200":
          description: Оборудование успешно удалено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Оборудование не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions:
    post:
      summary: Создание новой транзакции
      description: Регистрирует новую транзакцию (приход, расход, возврат, корректировка)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Транзакция успешно создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryTransaction"
        "400":
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Оборудование не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      summary: Получение списка транзакций
      description: Возвращает список всех транзакций с возможностью фильтрации
      security:
        - bearerAuth: []
      parameters:
        - name: item_id
          in: query
          description: Фильтр по идентификатору оборудования
          schema:
            type: string
        - name: transaction_type
          in: query
          description: Фильтр по типу транзакции
          schema:
            type: string
            enum: [intake, issue, return, adjustment]
        - name: responsible_user
          in: query
          description: Фильтр по ответственному пользователю
          schema:
            type: string
        - name: limit
          in: query
          description: Ограничение количества результатов
          schema:
            type: integer
            default: 50
        - name: skip
          in: query
          description: Количество пропускаемых результатов (для пагинации)
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Список транзакций
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryTransaction"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/item/{item_id}:
    get:
      summary: Получение транзакций по конкретному оборудованию
      description: Возвращает список всех транзакций для указанного оборудования
      security:
        - bearerAuth: []
      parameters:
        - name: item_id
          in: path
          required: true
          description: Идентификатор оборудования
          schema:
            type: string
      responses:
        "200":
          description: Список транзакций для оборудования
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryTransaction"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Оборудование не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invoices:
    post:
      summary: Создание новой накладной
      description: Создает новую накладную на основе транзакций
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Накладная успешно создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "400":
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      summary: Получение списка накладных
      description: Возвращает список всех накладных
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Список накладных
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invoices/{id}:
    get:
      summary: Получение информации о накладной
      description: Возвращает детальную информацию о конкретной накладной
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор накладной
          schema:
            type: string
      responses:
        "200":
          description: Информация о накладной
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Накладная не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/without-invoices:
    get:
      summary: Получение транзакций без накладных
      description: Возвращает список транзакций, для которых еще не созданы накладные
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Список транзакций без накладных
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryTransaction"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"